+++ CHANGELOG.md
@@ -1,5 +1,3 @@
 # HidingBar - Classic
 
-* Added option "Distance to bar border".
-* Added option "Distance between buttons".
-* Updated translations.
+* Fixed button scaling.